Refactor to run on window location change
The previous commit's code would run when netflix was initially loaded,
but would not run for subsequent page navigations. This update makes it
so the extension will look for the billboard row when the user navigates
to any site with a location of netflix.com/browse... and makes it so it
doesn't run on initial page load 😂.

Todo: run extension script on initial page load, and figure out how to
wait for any netflix loading events to finish before removing billboard.